+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 1. modify componer.json: make preview and test entry files more controllable
 2. update preview task and test task
+3. use sleep to fix preview bug when vendors created async